% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolkien_palette.R
\name{scale_color_tolkien}
\alias{scale_color_tolkien}
\alias{scale_colour_tolkien}
\alias{scale_colour_tolkien_d}
\alias{scale_color_tolkien_d}
\alias{scale_fill_tolkien_d}
\alias{scale_fill_tolkien}
\title{Talkien colour scales}
\usage{
scale_color_tolkien(option = "Sauron", ..., alpha = 1, begin = 0,
  end = 1, direction = 1, discrete = FALSE, house = NULL)

scale_colour_tolkien(option = "Sauron", ..., alpha = 1, begin = 0,
  end = 1, direction = 1, discrete = FALSE, house = NULL)

scale_colour_tolkien_d(option = "Sauron", ..., alpha = 1, begin = 0,
  end = 1, direction = 1)

scale_color_tolkien_d(option = "Sauron", ..., alpha = 1, begin = 0,
  end = 1, direction = 1)

scale_fill_tolkien_d(option = "Sauron", ..., alpha = 1, begin = 0,
  end = 1, direction = 1)

scale_fill_tolkien(option = "Sauron", ..., alpha = 1, begin = 0,
  end = 1, direction = 1, discrete = FALSE, house = NULL)
}
\arguments{
\item{option}{A character string indicating the colourmap  to use.}

\item{...}{parameters to \code{discrete_scale} or \code{scale_fill_gradientn}}

\item{alpha}{pass through parameter to \code{tolkien}}

\item{begin}{The (corrected) hue in [0,1] at which the tolkien colormap begins.}

\item{end}{The (corrected) hue in [0,1] at which the tolkien colormap ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors
are as output by \code{tolkien_pal}. If -1, the order of colors is reversed.}

\item{discrete}{generate a discrete palette? (default: \code{FALSE} - generate continuous palette)}

\item{house}{A character string indicating the colourmap from a option to use. This parameter is deprectaed, 'option' should be used instead.}
}
\description{
Uses the Tolkien color scale.
}
\details{
For \code{discrete == FALSE} (the default) all other arguments are as to
\link[ggplot2]{scale_fill_gradientn} or \link[ggplot2]{scale_color_gradientn}.
Otherwise the function will return a \code{discrete_scale} with the plot-computed
number of colors.
}
\examples{
library(ggplot2)



ggplot(mtcars, aes(factor(cyl), fill=factor(vs))) +
geom_bar() +
scale_fill_tolkien(discrete = TRUE, option = "Sauron")

ggplot(mtcars, aes(factor(gear), fill=factor(carb))) +
geom_bar() +
scale_fill_tolkien(discrete = TRUE, option = "Sauron")

ggplot(mtcars, aes(x = mpg, y = disp, colour = hp)) +
geom_point(size = 2) +
scale_colour_tolkien(option = "Sauron")




}
\author{
Alejandro Jiménez Rico \email{aljrico@gmail.com}
}
