% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolkien_palette.R
\name{tolkien_palettes}
\alias{tolkien_palettes}
\title{Available Palettes.}
\usage{
tolkien_palettes()
}
\value{
\code{tolkien_palettes} returns a character vector with the names of the palettes available to use.
}
\description{
This function outpus a vector containing the names of all the available palettes in the 'tolkien' package.
}
\examples{
tolkien_palettes()

}
\author{
Alejandro Jiménez Rico \email{aljrico@gmail.com}
}
